Run_MKClass_pirallel
The code that actually contains the parallel processing portion of the program. Start_mkclass_pirallel calls the program to use with the proper mpiexec information, so that the user doesn't need to understand how to do so for it to run. Must be on all nodes before running start_mkclass_pirallel or mpi will fail.